Contents

Contents

Index

PC-SAFT critical point

PC-SAFT saturation volumes

Index

EoSSuperancillaries.pcsaft_rhoc

EoSSuperancillaries.pcsaft_rholsat_reduced

EoSSuperancillaries.pcsaft_rhosat

EoSSuperancillaries.pcsaft_rhosat_reduced

EoSSuperancillaries.pcsaft_rhovsat_reduced

EoSSuperancillaries.pcsaft_tc

EoSSuperancillaries.pcsaft_theta

EoSSuperancillaries.pcsaft_vc

EoSSuperancillaries.pcsaft_vlsat

EoSSuperancillaries.pcsaft_vsat

EoSSuperancillaries.pcsaft_vvsat

PC-SAFT critical point

EoSSuperancillaries.pcsaft_tc — Function

Tc = pcsaft_tc(m,ϵ)

Returns the critical temperature of the PCSAFT equation of state.

Inputs:

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

Outputs:

Tc : Critical Temperature [K]. Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64).

Examples

julia> m, ϵ = 1.0, 150.03 #values for methane
(1.0, 150.03)

julia> Tc = pcsaft_tc(m, ϵ)
191.40058128833536

EoSSuperancillaries.pcsaft_rhoc — Function

rhoc = pcsaft_rhoc(m,σ)

Returns the critical density of the PCSAFT equation of state.

Inputs:

m: Segment length (no units)

σ: Monomer diameter [m]

Outputs:

rhoc : Critical density [mol/m^3]. Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64).

EoSSuperancillaries.pcsaft_vc — Function

vc = pcsaft_vc(m,σ)

Returns the critical volume of the PCSAFT equation of state. Equal to 1/pcsaft_rhoc(m,ϵ,σ)

Inputs:

m: Segment length (no units)

σ: Monomer diameter [m]

Outputs:

vc : Critical volume [m^3/mol]. Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64).

PC-SAFT saturation volumes

EoSSuperancillaries.pcsaft_rhosat — Function

rhol,rhov = pcsaft_rhosat(T,m,ϵ,σ)

Returns the saturation densities of the PCSAFT equation of state at the input temperature T.

Inputs:

T: Saturation temperature (Kelvin)

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

rhol : saturation liquid density [mol/m^3].

rhov : saturation vapour density [mol/m^3].

Returns NaN,NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and Tmin < T < Tc).

EoSSuperancillaries.pcsaft_vsat — Function

vl,vv = pcsaft_vsat(T,m,ϵ,σ)

Returns the saturation volumes of the PCSAFT equation of state at the input temperature T.

Inputs:

T: Saturation temperature (Kelvin)

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

vl : saturation liquid volume [m^3/mol].

vv : saturation vapour volume [m^3/mol].

Returns NaN,NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and Tmin < T < Tc).

EoSSuperancillaries.pcsaft_vlsat — Function

vv = pcsaft_vvsat(T,m,ϵ,σ)

Returns the saturation vapour volume of the PCSAFT equation of state at the input temperature T.

Inputs:

T: Saturation temperature (Kelvin)

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

vl : saturation liquid volume [m^3/mol].

Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and Tmin < T < Tc).

EoSSuperancillaries.pcsaft_vvsat — Function

vl = pcsaft_vlsat(T,m,ϵ,σ)

Returns the saturation liquid volume of the PCSAFT equation of state at the input temperature T.

Inputs:

T: Saturation temperature (Kelvin)

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

vl : saturation liquid volume [m^3/mol].

Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and Tmin < T < Tc).

EoSSuperancillaries.pcsaft_theta — Function

Θ,status = pcsaft_theta(T̃,m)
Θ,status = pcsaft_theta(T̃,m,T̃c)

Calculates reduced scaled temperature parameter used in PCSAFT saturation volume superancillary

Inputs:

T̃: Temperature divided by reduced interaction energy (T/ϵ) (no units)

m: Segment length (no units)

T̃c: (Optional) Reduced critical temperature (Tc/ϵ) (no units)

Outputs:

Θ : scaled temperature parameter (no units)

status : Symbol used to signal if the combination of T̃,m is valid or not. It can return one of the following:

:inrange : if the combination is valid

:nonfinite : if any input is not finite

:below_mmin : if m < 1.0

:over_mmax : if m > 64.0

:below_Tmin : if T̃ < T̃min, where T̃min = T̃c*exp(-2.20078778)*m^0.37627892

:over_Tmax : if T̃ > T̃c

EoSSuperancillaries.pcsaft_rhosat_reduced — Function

ρ̃l,ρ̃v = pcsaft_rhosat_reduced(Θ,m)

Calculates the reduced saturation densities. The actual densities are calculated as ρᵢ = ρ̃ᵢ/(N_A*σ^3)

Inputs:

Θ: reduced temperature parameter (no units)

m: Segment length (no units)

Outputs:

ρ̃l : reduced saturation liquid density (no units)

ρ̃v : reduced saturation vapour density (no units)

Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and 0 < Θ < 1).

EoSSuperancillaries.pcsaft_rholsat_reduced — Function

ρ̃l = pcsaft_rholsat_reduced(Θ,m)

Calculates the reduced liquid saturation density. The actual density is ρ = ρ̃/(N_A*σ^3)

Inputs:

Θ: reduced temperature parameter (no units)

m: Segment length (no units)

Outputs:

ρ̃l : reduced saturation liquid density (no units)

Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and 0 < Θ < 1).

EoSSuperancillaries.pcsaft_rhovsat_reduced — Function

ρ̃v = pcsaft_rhovsat_reduced(Θ,m)

Calculates the reduced vapour saturation density. The actual density is ρ = ρ̃/(N_A*σ^3)

Inputs:

Θ: reduced temperature parameter (no units)

m: Segment length (no units)

Outputs:

ρ̃v : reduced saturation vapour density (no units)

Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and 0 < Θ < 1).