Contents

Contents

Index

Redlich-Kwong superancillaries

Index

EoSSuperancillaries.rk_psat

EoSSuperancillaries.rk_vlsat

EoSSuperancillaries.rk_vsat

EoSSuperancillaries.rk_vvsat

Redlich-Kwong superancillaries

EoSSuperancillaries.rk_vsat — Function

vl,vv = rk_vsat(T,a,b)

Returns the saturation liquid and vapour volumes of the Redlich-Kwong cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

vl : Saturation Liquid Volume [m^3]. Returns NaN if the value is outside the range of the ancillary.

vv : Saturation Vapour Volume [m^3]. Returns NaN if the value is outside the range of the ancillary.

EoSSuperancillaries.rk_vlsat — Function

vl = rk_vlsat(T,a,b)

Returns the saturation liquid volume of the Redlich-Kwong cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

vl : Saturation Liquid Volume [m^3]. Returns NaN if the value is outside the range of the ancillary.

EoSSuperancillaries.rk_vvsat — Function

vv = rk_vvsat(T,a,b)

Returns the saturation vapour volume of the Redlich-Kwong cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

vv : Saturation Vapour Volume [m^3]. Returns NaN if the value is outside the range of the ancillary.

EoSSuperancillaries.rk_psat — Function

p = rk_psat(T,a,b)

Returns the saturation pressure of the Redlich-Kwong cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

p : Saturation pressure [Pa]. Returns NaN if the value is outside the range of the ancillary.