Contents

Index

van der Wals superancillaries

EoSSuperancillaries.vdw_vlsat — Function

vl = vdw_vlsat(T,a,b)

Returns the saturation liquid volume of the vdW cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

vl : Saturation Liquid Volume [m^3]. Returns NaN if the value is outside the range of the ancillary.

EoSSuperancillaries.vdw_vvsat — Function

vv = vdw_vvsat(T,a,b)

Returns the saturation vapour volume of the vdW cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

vv : Saturation Vapour Volume [m^3]. Returns NaN if the value is outside the range of the ancillary.

EoSSuperancillaries.vdw_psat — Function

p = vdw_psat(T,a,b)

Returns the saturation pressure of the vdW cubic equation of state.

Inputs:

T: Temperature (Kelvin)

a: Atraction parameter [m^6*Pa/mol]

b: Covolume [m^3/mol]

Outputs:

p : Saturation pressure [Pa]. Returns NaN if the value is outside the range of the ancillary.