Contents

Index

PC-SAFT superancillaries

EoSSuperancillaries.pcsaft_tc — Function

Tc = pcsaft_tc(m,ϵ)

Returns the critical temperature of the PCSAFT equation of state.

Inputs:

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

Outputs:

Tc : Critical Temperature [K]. Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64).

Examples

julia> m, ϵ = 1.0, 150.03 #values for methane
(1.0, 150.03)

julia> Tc = pcsaft_tc(m, ϵ)
191.40058128833536

EoSSuperancillaries.pcsaft_rhoc — Function

rhoc = pcsaft_rhoc(m,ϵ,σ)

Returns the critical density of the PCSAFT equation of state.

Inputs:

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

rhoc : Critical density [mol/m^3]. Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64).

EoSSuperancillaries.pcsaft_vc — Function

vc = pcsaft_vc(m,ϵ,σ)

Returns the critical volume of the PCSAFT equation of state. Equal to 1/pcsaft_rhoc(m,ϵ,σ)

Inputs:

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

vc : Critical volume [m^3/mol]. Returns NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64).

EoSSuperancillaries.pcsaft_rhosat — Function

vc = pcsaft_rhosat(T,m,ϵ,σ)

Returns the saturation volumes of the PCSAFT equation of state at the input temperature T.

Inputs:

T: Saturation temperature (Kelvin)

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

rhol : saturation liquid volume [m^3/mol].

rhov : saturation vapour volume [m^3/mol].

Returns NaN,NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and Tmin < T < Tc).

EoSSuperancillaries.pcsaft_vsat — Function

vc = pcsaft_vsat(T,m,ϵ,σ)

Returns the saturation volume of the PCSAFT equation of state at the input temperature T.

Inputs:

T: Saturation temperature (Kelvin)

m: Segment length (no units)

ϵ: Reduced interaction energy [K]

σ: Monomer diameter [m]

Outputs:

vl : saturation liquid volume [m^3/mol].

vv : saturation vapour volume [m^3/mol].

Returns NaN,NaN if the value is outside the range of the ancillary (1 ≤ m ≤ 64 and Tmin < T < Tc).